+++ app/src/settings_tabs/advanced_settings/OverlaySettings.css
@@ -1,0 +1,31 @@
+.dcc-accordion-wrapper {
+    min-width: 10vw;
+    width: 30vw;
+    max-width: 70vw;
+    overflow: auto;
+    border-radius: 4px;
+    z-index: 0;
+    padding-left: 10px;
+    height: 20vh;
+    resize: horizontal;
+  }
+  
+  .MuiTypography-root.dcc-not-allowed {
+    color: gray;
+    font-size: 0.97rem;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    cursor: default;
+  }
+  
+  .dcc-column-leaf {
+    margin-left: 23.2px;
+    cursor: pointer;
+  }
+  
+  .dcc-column-leaf-not-allowed {
+    margin-left: 23.2px;
+    cursor: default;
+  }
+
